Add initial demo implementation of new execution model
The intended demo currently does not work: it loops infinitely.
Right now we cannot ever have loops in wire graphs, because that leads
to infinite processing (since writing to an input means triggering
processing, even if the value written is equal to whatever is already
in there).

We should consider:
  - implementing limits on how long processing can take (so that
    infinite loops are aborted eventually)
  - seeing if we can do something smart to allow loops that reach
    a fixed point